requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
langchain
langchain-openai